fix: move timespan function to data.py
